Complete C90 compliance and build integration fixes
Additional C90 compliance fixes:
- orvos_tiditem.c: Cast nelements to uint64 in comparisons to fix sign warnings
- orvos_tupslot.c: Add (void) statements for unused parameters
- orvos_wal.c: Add (void) statement for unused blkno parameter

Build system:
- src/backend/access/Makefile: Enable orvos unconditionally (moved into SUBDIRS list)
  Previously was conditionally included based on with_orvos variable

Documentation:
- README.md: Updated with project documentation
- .local-gitignore: Added AUTHORS_NOTE.md to ignore list

These changes ensure clean compilation with zero warnings and proper
build system integration.

+++ src/backend/access/orvos/orvos_tiditem.c
@@ -96,7 +96,7 @@ ovbt_tid_item_create_for_range(ovtid tid, int nelements, OVUndoRecPtr undo_ptr)
 	}
 
 	total_encoded = 0;
-	while (total_encoded < nelements)
+	while (total_encoded < (uint64) nelements)
 	{
 		OVTidArrayItem *newitem;
 		Size		itemsz;
@@ -119,7 +119,7 @@ ovbt_tid_item_create_for_range(ovtid tid, int nelements, OVUndoRecPtr undo_ptr)
 		second_delta = 1;
 		num_tids = 0;
 		for (num_codewords = 0;
-			 num_codewords < OVBT_MAX_ITEM_CODEWORDS && total_encoded < nelements && num_tids < OVBT_MAX_ITEM_TIDS;
+			 num_codewords < OVBT_MAX_ITEM_CODEWORDS && total_encoded < (uint64) nelements && num_tids < OVBT_MAX_ITEM_TIDS;
 			 num_codewords++)
 		{
 			uint64		codeword;
